Fix external subtitles not activating if transcoding
